backend/pyproject.toml
[project]
name = "codeproof-backend"
version = "0.1.0"
description = "AI Code Review SaaS Platform - Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.26.0",
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",
    "qdrant-client>=1.7.0",
    "openai>=1.10.0",
    "google-genai>=0.3.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
    "tree-sitter-php>=0.22.0",
    "python-multipart>=0.0.6",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.7",
    "numpy>=1.26.0",
    "email-validator (>=2.3.0,<3.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    # Legacy/unused services (not integrated yet)
    "app/services/clone_service.py",
    "app/services/codebase_doc_service.py",
    "app/services/compliance_service.py",
    "app/services/deep_analysis_service.py",
    "app/services/parser_service.py",
    # Test files
    "tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
